Update vendored dep
github.com/yosssi/gmq@v0.0.1 contained a race/bug that
could result in a deadlock after recieving a ping.

While further debugging the lock up/leak I confirmed that
it was caused by a ping timing condition and discovered
that go mod had selected an older commit that was tagged
as a release.
Updating to latest master should resolve the issue.